Parallelize e2e tests and reduce polling times
